Yikes!  Too many files!!
Fixing #220 ; it seems that the problem was the way fault_codes were being set.  The value returned by hash() on a 64-bit machine was evidently too large for something or other in XML RPC.  To fix it, nearly all exceptions have been moved to a Cobalt.Exceptions module, where they can use a shared object to automatically set fault_codes to something unique.

git-svn-id: https://svn.mcs.anl.gov/repos/cobalt/trunk@1221 c55eebd0-760c-0410-adae-a6e4ca644001