fixes #433
The client now no longer restricts the number range of minutes or hours.  
This does get parsed correctly by the submission script, and qstat shows a 
a correct time, even if the user does something truly strange like a time of 
01:90:00.  Moreover the case of 00:60:00 now acts as requested.

The script still rejects negative values for days, hours, minutes or seconds.



git-svn-id: https://svn.mcs.anl.gov/repos/cobalt/trunk@1790 c55eebd0-760c-0410-adae-a6e4ca644001